@@ -3,12 +3,22 @@ import {WebSocketSubject} from "rxjs/webSocket";
 import {Subject} from "rxjs";
 import {HttpClient} from "@angular/common/http";
 
+export class ResponseMessage {
+  constructor(
+    public type: string,
+    public retroResults: RetroResults,
+    public trueConfessions: string[]
+  ) {
+  }
+}
+
 export class RetroResults {
   constructor(
     public numResults: number,
     public tempResults: AggregateResults,
     public safetyResults: AggregateResults
-  ) { }
+  ) {
+  }
 }
 
 export class AggregateResults {
@@ -16,7 +26,8 @@ export class AggregateResults {
     public min: number,
     public max: number,
     public avg: number,
-  ) { }
+  ) {
+  }
 }
 
 @Injectable({
@@ -25,35 +36,58 @@ export class AggregateResults {
 export class AppService {
 
   results: Subject<RetroResults>;
+  trueConfessions: Subject<string[]>;
 
   private socket: WebSocketSubject<object>;
+
   constructor(private http: HttpClient) {
     this.results = new Subject<RetroResults>();
+    this.trueConfessions = new Subject<string[]>();
+
     const protocol = window.location.protocol === 'https:' ? 'wss:' : 'ws:';
     const url = `${protocol}//${window.location.host}/ws`;
     this.socket = new WebSocketSubject<object>(url);
     this.listen();
   }
 
-  submit(temp: number, safety: number) {
+  submitTempSafety(temp: number, safety: number) {
     const msg = {
-      temp: temp,
-      safety: safety
+      type: 'tempSafety',
+      tempSafety: {
+        temp: temp,
+        safety: safety
+      }
+    };
+
+    this.socket.next(msg);
+  }
+
+  submitTrueConfession(trueConfession: string) {
+    const msg = {
+      type: 'trueConfession',
+      trueConfession: trueConfession
     };
 
     this.socket.next(msg);
   }
 
   listen() {
-    this.socket.subscribe(msg => {
-      this.results.next(msg as RetroResults);
+    this.socket.subscribe((msg: ResponseMessage) => {
+      switch (msg.type) {
+        case "retroResults":
+          this.results.next(msg.retroResults);
+          break;
+        case "trueConfession":
+          this.trueConfessions.next(msg.trueConfessions);
+          break;
+      }
     }, err => {
       console.error(err);
     });
   }
 
   clear() {
-    this.http.post('clear', null).subscribe( () => {
+    this.http.post('clear', null).subscribe(() => {
     });
   }
 }
